move approx root and implement naive monotonic sqrt
