+++ packages/math/src/pool/concentrated/math.ts
@@ -1,6 +1,5 @@
 import { Dec, Int } from "@keplr-wallet/unit";
 
-import { smallestDec } from "./const";
 import { tickToSqrtPrice } from "./tick";
 
 /** The `@keplr-wallet/unit` `Dec` object doesn't have the `mulRoundUp()` function
@@ -222,52 +221,6 @@ export function getFeeChargePerSwapStepOutGivenIn(
   }
 
   return feeChargeTotal;
-}
-
-export function approxSqrt(dec: Dec, maxIters = 300): Dec {
-  return approxRoot(dec, 2, maxIters);
-}
-
-/**
-  Approximate root using Newton's method.
-
-  This function approximates the square root of a given decimal.
-  It uses Newton's method to approximate the square root.
-  It does this by iterating through the formula:
-  x_{n+1} = x_n - (x_n^2 - a) / (2 * x_n)
-  where x_0 is the initial approximation, a is the number whose
-  square root we want to find, and x_n is the current approximation.
-  The number of iterations is controlled by maxIters.
-
-   TODO: move to decimal object see: https://github.com/chainapsis/keplr-wallet/pull/674
- */
-export function approxRoot(dec: Dec, root: number, maxIters = 300): Dec {
-  if (dec.isNegative()) {
-    return approxRoot(dec.neg(), root).neg();
-  }
-
-  if (root === 1 || dec.isZero() || dec.equals(new Dec(1))) {
-    return dec;
-  }
-
-  if (root === 0) {
-    return new Dec(1);
-  }
-
-  let [guess, delta] = [new Dec(1), new Dec(1)];
-  for (let i = 0; delta.abs().gt(smallestDec) && i < maxIters; i++) {
-    let prev = guess.pow(new Int(root - 1));
-    if (prev.isZero()) {
-      prev = smallestDec;
-    }
-    delta = dec.quo(prev);
-    delta = delta.sub(guess);
-    delta = delta.quoTruncate(new Dec(root));
-
-    guess = guess.add(delta);
-  }
-
-  return guess;
 }
 
 // https://github.com/osmosis-labs/osmosis/blob/1c5f166d180ca6ffdd0a4068b97422c5c169240c/x/concentrated-liquidity/math/math.go#L163
