+++ packages/math/src/utils.ts
@@ -1,5 +1,7 @@
 import { Dec, Int } from "@keplr-wallet/unit";
 
+import { smallestDec } from "./pool/concentrated/const";
+
 const powPrecision = new Dec("0.00000001");
 
 // This file is intended to be a typescript implementation of power matching
@@ -97,4 +99,65 @@ function powInt(base: Dec, power: Int): Dec {
   }
 
   return base.mul(tmp);
+}
+
+/**
+  Approximate root using Newton's method.
+
+  This function approximates the square root of a given decimal.
+  It uses Newton's method to approximate the square root.
+  It does this by iterating through the formula:
+  x_{n+1} = x_n - (x_n^2 - a) / (2 * x_n)
+  where x_0 is the initial approximation, a is the number whose
+  square root we want to find, and x_n is the current approximation.
+  The number of iterations is controlled by maxIters.
+
+   TODO: move to decimal object see: https://github.com/chainapsis/keplr-wallet/pull/674
+ */
+export function approxRoot(dec: Dec, root: number, maxIters = 300): Dec {
+  if (dec.isNegative()) {
+    return approxRoot(dec.neg(), root).neg();
+  }
+
+  if (root === 1 || dec.isZero() || dec.equals(new Dec(1))) {
+    return dec;
+  }
+
+  if (root === 0) {
+    return new Dec(1);
+  }
+
+  let [guess, delta] = [new Dec(1), new Dec(1)];
+  for (let i = 0; delta.abs().gt(smallestDec) && i < maxIters; i++) {
+    let prev = guess.pow(new Int(root - 1));
+    if (prev.isZero()) {
+      prev = smallestDec;
+    }
+    delta = dec.quo(prev);
+    delta = delta.sub(guess);
+    delta = delta.quoTruncate(new Dec(root));
+
+    guess = guess.add(delta);
+  }
+
+  return guess;
+}
+
+export function approxSqrt(dec: Dec, maxIters = 300): Dec {
+  return approxRoot(dec, 2, maxIters);
+}
+
+// monotonicSqrt does a naive 1 ULP shift to the output of approxSqrt to approximate monotonicity.
+// While this function is not technically monotonic, it is close enough for our purposes.
+// True monotonicity can be achieved by replacing the `approxRoot` with a call to big int's lower level
+// square root function (floor sqrt), which in conjunction with the 1 ULP shift will guarantee monotonicity.
+export function monotonicSqrt(dec: Dec, maxIters = 300): Dec {
+  let estimate = approxRoot(dec, 2, maxIters);
+  const inverseValue = estimate.pow(new Int(2));
+
+  if (inverseValue.lt(estimate)) {
+    estimate = estimate.add(smallestDec);
+  }
+
+  return estimate;
 }
